repeated all the work.. with final data
Read in the data, modified the script for normalizing AND smoothing the data, including the WV. I updated the epsilons range windows, and I converted to COV vs TIME. I played around with splitting frequencies and fixed the weights. I created images for all of these and put them in a presentation. I had issue with the splititng region but I think it's good now.