got Ea with cov also testing temperatures, 480 seems to dip
I tested fitting too. Still issue but I fixed the x_A fraction function. I need to clean up that code and double check all the regions that are used for the REGIONS.
Additionally, it might be worth to explore the adsorption part using k_oA because it seems strange.